feat: wrap in suspense
